requirements.txt
amqp==2.2.1
billiard==3.5.0.3
boto==2.43.0
celery==4.0.2
Django==1.11.4
django-mongodb-engine==0.6.0
django-rest-framework-mongoengine==3.3.1
djangorestframework==3.6.4
djangotoolbox==1.8.0
kombu==4.1.0
mongoengine==0.13.0
pymongo==3.5.1
python-json-logger==0.1.8
PyQRCode==1.2.1
pytz==2017.2
redis==2.10.6
redisbeat==1.1.0
six==1.10.0
vine==1.1.4
